+++ unstructured_ingest/v2/pipeline/steps/stage.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Callable, Optional, TypedDict
 
-from unstructured_ingest.v2.interfaces.file_data import FileData
+from unstructured_ingest.v2.interfaces.file_data import file_data_from_file
 from unstructured_ingest.v2.interfaces.upload_stager import UploadStager
 from unstructured_ingest.v2.logger import logger
 from unstructured_ingest.v2.pipeline.interfaces import PipelineStep
@@ -39,11 +39,13 @@ class UploadStageStep(PipelineStep):
         self, fn: Callable, path: str, file_data_path: str
     ) -> UploadStageStepResponse:
         path = Path(path)
+        # Maintain extension
+        output_filename = f"{self.get_hash(extras=[path.name])}{path.suffix}"
         fn_kwargs = {
             "elements_filepath": path,
-            "file_data": FileData.from_file(path=file_data_path),
+            "file_data": file_data_from_file(path=file_data_path),
             "output_dir": self.cache_dir,
-            "output_filename": self.get_hash(extras=[path.name]),
+            "output_filename": output_filename,
         }
         if not asyncio.iscoroutinefunction(fn):
             staged_output_path = fn(**fn_kwargs)